Use dispatch URI scheme from kolibri_app.config
Instead of hard-coding `x-kolibri-dispatch://`, we'll use the URI scheme
specified by `DISPATCH_URI_SCHEME` in `kolibri_app.config`. This fixes
the plugin generating incorrect launchers inside custom builds of the
Kolibri desktop app. For simplicity, we assume that Kolibri itself is
running in the same Python environment as the desktop app.

https://phabricator.endlessm.com/T35614